K_estimated
A Jupyter Notebook is an interactive computing environment that allows you to create and share documents containing live code, equations, visualizations, and explanatory text. In this Jupyter Notebook, I load well log data and propose a function to calculate the thermal conductivity of the well based on the volumes of clay and sand, along with their respective conductivity values. The function takes into account the properties of different components and utilizes the porosity information from the well log data. By applying the function, we can estimate the thermal conductivity profile along the depth of the well.